refactor(graph): let the classifier own metadata resolution too
readGraphArtifactMetadata was the third implementation of "which
artifact does this path mean?", after loadGraph and
loadGraphArtifactFromPath. It could not call the classifier because the
classifier imports this module.

Broken by moving the on-disk markers -- the v2 header, the tombstone,
the moved prefix and its predicate -- into a leaf module both can
import. graph-artifact.ts re-exports them, so its many importers are
unaffected, and it can now depend on the classifier without a cycle.

Metadata resolution for a conventional legacy path now goes through
classifyWorkspaceGraph, matching the loader exactly. The remaining
inline marker check is reached only for a moved marker outside out/,
where directory classification does not apply.

Mutation sweep still 19/19.
